Move currency exceptions to domain
