+++ backend/src/main/java/com/alligator/market/backend/instrument_catalog/currency_pair/web/PairExceptionHandler.java
@@ -2,10 +2,10 @@ package com.alligator.market.backend.instrument_catalog.currency_pair.web;
 
 import com.alligator.market.backend.common.web.ApiResponse;
 import com.alligator.market.backend.common.web.ResponseEntityFactory;
-import com.alligator.market.backend.instrument_catalog.currency_pair.exception.CurrencyFromPairNotFoundException;
-import com.alligator.market.backend.instrument_catalog.currency_pair.exception.DuplicatePairException;
-import com.alligator.market.backend.instrument_catalog.currency_pair.exception.PairNotFoundException;
-import com.alligator.market.backend.instrument_catalog.currency_pair.exception.EqualCurrenciesInPairException;
+import com.alligator.market.domain.instrument.currency_pair.exception.CurrencyFromPairNotFoundException;
+import com.alligator.market.domain.instrument.currency_pair.exception.DuplicatePairException;
+import com.alligator.market.domain.instrument.currency_pair.exception.PairNotFoundException;
+import com.alligator.market.domain.instrument.currency_pair.exception.EqualCurrenciesInPairException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
