+++ domain/src/main/java/com/alligator/market/domain/instrument/currency/exception/CurrencyUsedInPairsException.java
@@ -1,4 +1,4 @@
-package com.alligator.market.backend.instrument_catalog.currency.exception;
+package com.alligator.market.domain.instrument.currency.exception;
 
 /**
  * Нельзя удалить валюту, если она используется в валютных парах.
