+++ domain/src/main/java/com/alligator/market/domain/instrument/currency_pair/exception/PairNotFoundException.java
@@ -1,4 +1,4 @@
-package com.alligator.market.backend.instrument_catalog.currency_pair.exception;
+package com.alligator.market.domain.instrument.currency_pair.exception;
 
 import jakarta.persistence.EntityNotFoundException;
